Refactor wcsmod module into individual modules for each wrapper
- puts each wrapper in a separate module
- for easier administration and editing
